feat(share-ui): split android and ios share screens
